remove unused global variables

+++ src/Scope.cpp
@@ -10,7 +10,7 @@ template <typename T> struct ScopeBase {
     const_iterator begin() const { return data.data(); }
     iterator end() { return data.data() + data.size(); }
     const_iterator end() const { return data.data() + data.size(); }
-    T *getSym(IdentRef sym, size_t &idx) {
+    T *getSym(IdentRef sym, unsigned &idx) {
         for (size_t i = data.size(); i--;) {
             Storage &elem = data[i];
             if (elem.sym == sym)
@@ -26,11 +26,11 @@ template <typename T> struct ScopeBase {
         }
         return nullptr;
     }
-    T &getSym(size_t index) { return data[index].info; }
-    Storage &getSymFull(size_t index) { return data[index]; }
-    StringRef getSymName(size_t index) { return data[index].sym->getKey(); }
-    size_t putSym(IdentRef sym, const T &Elt) {
-        size_t size = data.size();
+    T &getSym(unsigned index) { return data[index].info; }
+    Storage &getSymFull(unsigned index) { return data[index]; }
+    StringRef getSymName(unsigned index) { return data[index].sym->getKey(); }
+    unsigned putSym(IdentRef sym, const T &Elt) {
+        unsigned size = data.size();
         Storage s;
         s.sym = sym;
         s.info = Elt;
@@ -49,26 +49,26 @@ template <typename T> struct ScopeBase {
 template <typename T> struct BlockScope : public ScopeBase<T> {
     // https://stackoverflow.com/q/1120833
     // Derived template-class access to base-class member-data
-    SmallVector<size_t, 8> blocks{};
-    size_t maxSyms = 0;
-    size_t numSymsThisBlock() const { return blocks.back(); }
+    SmallVector<unsigned, 8> blocks{};
+    unsigned maxSyms = 0;
+    unsigned numSymsThisBlock() const { return blocks.back(); }
     void push() { blocks.push_back(this->data.size()); }
     void pop() {
         assert(blocks.size() && "mismatched push/pop: no stack to pop()!");
-        maxSyms = std::max(maxSyms, this->data.size());
+        maxSyms = std::max(static_cast<size_t>(maxSyms), this->data.size());
         this->data.resize(blocks.back());
         blocks.pop_back();
     }
-    void finalizeGlobalScope() { maxSyms = std::max(maxSyms, this->data.size()); }
+    void finalizeGlobalScope() { maxSyms = std::max(static_cast<size_t>(maxSyms), this->data.size()); }
     bool isInGlobalScope(IdentRef Name) {
         assert(blocks.size() && "no global scope");
-        for (size_t i = 0; i < blocks.size(); i++) {
+        for (unsigned i = 0; i < blocks.size(); i++) {
             if (this->data[i].sym == Name)
                 return true;
         }
         return false;
     }
-    bool isInGlobalScope(size_t idx) {
+    bool isInGlobalScope(unsigned idx) {
         assert(blocks.size() && "no global scope");
         return idx < blocks.front();
     }
@@ -82,7 +82,7 @@ template <typename T> struct BlockScope : public ScopeBase<T> {
         }
         return nullptr;
     }
-    T *getSymInCurrentScope(IdentRef Name, size_t &idx) {
+    T *getSymInCurrentScope(IdentRef Name, unsigned &idx) {
         for (size_t i = numSymsThisBlock(); i < this->data.size(); ++i) {
             auto &elem = this->data[i];
             if (elem.sym == Name)
@@ -98,7 +98,7 @@ template <typename T> struct BlockScope : public ScopeBase<T> {
     }
 };
 template <typename T> struct FunctionAndBlockScope : public BlockScope<T> {
-    size_t _current_function_offset;
+    unsigned _current_function_offset;
     auto current_function() { return this->data.data() + _current_function_offset; }
     auto current_function() const { return this->data.data() + _current_function_offset; }
     void push_function() { _current_function_offset = this->data.size(); }
